requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
pyswip==0.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tzdata==2025.3
